# src/presentation/api/v1/input-schemas/user/create-user.input-schema.ts
import { MongooseBaseObjectSchema } from '@infrastructure/mongoose';
import { ApiProperty } from '@nestjs/swagger';

export class CreateUserInputSchema extends MongooseBaseObjectSchema {
  @ApiProperty({
    description: 'Tên bài viết',
    example: 'Hướng dẫn sử dụng NestJS',
  })
  name!: string;
}