# src/presentation/api/v1/input-schemas/user/update-user.input-schema.ts
import { MongooseBaseObjectSchema } from '@infrastructure/mongoose';
import { ApiProperty } from '@nestjs/swagger';

export class UpdateUserInputSchema extends MongooseBaseObjectSchema {
  @ApiProperty({
    description: 'Tên bài viết',
    example: 'Hướng dẫn sử dụng NestJS',
  })
  name!: string;
}